requirements.txt
aiohttp==3.7.4.post0
aiohttp-auth==0.1.1
aiohttp-socks==0.6.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
construct==2.8.8
Django==3.1.7
idna==2.10
ipaddress==1.0.16
multidict==5.1.0
pycares==3.1.1
pycparser==2.20
PySocks==1.7.1
python-socks==1.2.3
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
six==1.15.0
speedtest-cli==2.1.3
sqlparse==0.4.1
ticket-auth==0.1.4
transitions==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.3
